Solution settings class

Solution settings class

Solution settings class

Details

Definition for the SolutionSettings class.

The argument to value should be a list with the following elements:

id

character identifier for theme or weight.

parameter

character name of parameter. Available options are: "status", "factor", or "goal".

type

character indicating if the parameter is for a theme or weight. Available options are: "theme" or "weight".

The argument to value should be a list with the following elements:

id

character identifier for theme or weight.

parameter

character name of parameter. Available options are: "status", "factor", or "goal".

value

numeric or logical value for new parameter.

type

character indicating if the parameter is for a theme or weight. Available options are: "theme" or "weight".

Public fields

theme_ids

character vector of identifiers for the themes.

weight_ids

character vector of identifiers for the weights.

themes

list of Theme objects.

weights

list of Weight objects.

Methods

Public methods

SolutionSettings$new()

SolutionSettings$print()

SolutionSettings$repr()

SolutionSettings$get_theme()

SolutionSettings$get_weight()

SolutionSettings$get_parameter()

SolutionSettings$set_parameter()

SolutionSettings$get_widget_data()

SolutionSettings$get_prioritization_data()

SolutionSettings$clone()

Method new()

Create a SolutionSettings object.

Usage

SolutionSettings$new(themes, weights)

Arguments

themes

list of Theme objects.

weights

list of Weight objects.

Returns

A new SolutionSettings object.

Method new()

Create a SolutionSettings object.

Usage

SolutionSettings$new(themes, weights)

Arguments

themes

list of Theme objects.

weights

list of Weight objects.

Returns

A new SolutionSettings object.

Method print()

Print the object.

Usage

SolutionSettings$print(...)

Arguments

...

not used.

Method print()

Print the object.

Usage

SolutionSettings$print(...)

Arguments

...

not used.

Method repr()

Generate a character summarizing the representation of the object.

Usage

SolutionSettings$repr()

Returns

character value.

Method repr()

Generate a character summarizing the representation of the object.

Usage

SolutionSettings$repr()

Returns

character value.

Method get_theme()

Get a theme.

Usage

SolutionSettings$get_theme(value)

Arguments

value

character theme identifier.

Returns

Theme object.

Method get_theme()

Get a theme.

Usage

SolutionSettings$get_theme(value)

Arguments

value

character theme identifier.

Returns

Theme object.

Method get_weight()

Get a weight.

Usage

SolutionSettings$get_weight(value)

Arguments

value

character weight identifier.

Returns

Weight object.

Method get_weight()

Get a weight.

Usage

SolutionSettings$get_weight(value)

Arguments

value

character weight identifier.

Returns

Weight object.

Method get_parameter()

Get a parameter for a weight or theme.

Usage

SolutionSettings$get_parameter(value)

Arguments

value

list with new parameter information (see Details section)

Method get_parameter()

Get a parameter for a weight or theme.

Usage

SolutionSettings$get_parameter(value)

Arguments

value

list with new parameter information (see Details section)

Method set_parameter()

Set a parameter for a weight or theme.

Usage

SolutionSettings$set_parameter(value)

Arguments

value

list with new parameter information (see Details section)

Method set_parameter()

Set a parameter for a weight or theme.

Usage

SolutionSettings$set_parameter(value)

Arguments

value

list with new parameter information (see Details section)

Method get_widget_data()

Get data for displaying the theme in a solutionSettings() widget.

Usage

SolutionSettings$get_widget_data()

Returns

list with widget data.

Method get_widget_data()

Get data for displaying the theme in a solutionSettings() widget.

Usage

SolutionSettings$get_widget_data()

Returns

list with widget data.

Method get_prioritization_data()

Get data for generating a prioritization.

Usage

SolutionSettings$get_prioritization_data()

Returns

list with data.

Method get_prioritization_data()

Get data for generating a prioritization.

Usage

SolutionSettings$get_prioritization_data()

Returns

list with data.

Method clone()

The objects of this class are cloneable with this method.

Usage

SolutionSettings$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

SolutionSettings$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.