New solution manager

Constructs a widget for generating new solutions and managing their associated settings.

newSolutionManager(
  themes,
  weights,
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

themes
list of Theme objects to include in the manager.

weights
list of Weight objects to include in the manager.

width
character width of the displayed widget. Defaults to NULL such that the widget is automatically sized.

height
character width of the displayed widget. Defaults to NULL such that the widget is automatically sized.

elementId
character HTML identifier for the widget. Defaults to NULL.

Server value

The widget sends the following values to the server (where elementId corresponds to the elementId):

elementId_run

character name to use for a new solution. This value is updated when the user clicks the button to generate a new solution.

elementId_setting

list containing updated settings. This value is updated when the user updates the settings (e.g. status, factor, goal) for a Weight or Theme. Specifically, the list contains the following elements:

id

character identifier for the theme or weight.

parameter

character name of the updated parameter. Available options include: "status", "factor", or "goal".

value

new numeric or logical values.

type

character indicating if the updated parameter corresponds to a theme or weight.

Examples

# TODO.