New dataset

Create a new Dataset object.

new_dataset(
  spatial_path,
  attribute_path,
  boundary_path,
  spatial_data = NULL,
  attribute_data = NULL,
  boundary_data = NULL,
  id = uuid::UUIDgenerate()
)

Arguments

spatial_path
character file path for spatial data.

attribute_path
character file path for attribute data.

boundary_path
character file path for boundary data.

spatial_data
NULL, sf::st_sf(), or raster::raster() object. Defaults to NULL such that data are automatically imported using the argument to spatial_path.

attribute_data
NULL, or tibble::tibble() object. Defaults to NULL such that data are automatically imported using the argument to attribute_path.

boundary_data
NULL, or Matrix::sparseMatrix() object. Defaults to NULL such that data are automatically imported using the argument to boundary_path.

id
character unique identifier. Defaults to a random identifier (uuid::UUIDgenerate()).

Value

A Dataset object.

Examples

# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# print object
print(d)
#> Dataset
#>   paths:
#>     spatial: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_spatial.tif
#>     attribute: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_attribute.csv.gz
#>     boundary: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_boundary.csv.gz