New feature

Create a new Feature object.

new_feature(
  name,
  layer,
  initial_status = TRUE,
  initial_goal = 0.1,
  min_goal = 0,
  max_goal = 1,
  step_goal = 0.01,
  limit_goal = 0.1,
  current_label = "Current",
  icon = "map-marked-alt",
  id = uuid::UUIDgenerate()
)

Arguments

name
character Name of the theme.

layer
Layer object.

initial_status
logical The initial status. This is used to display information on whether the theme is selected (or not) for subsequent analysis. Defaults to TRUE

initial_goal
numeric The initial goal for the feature. Note that goal values are specified as proportions, such that a value of 0.1 corresponds to 10%. Defaults to 0.3 (i.e. 30%).

min_goal
numeric The minimum goal (inclusive) shown for the feature. Note that goal values are specified as proportions, such that a value of 0.1 corresponds to 10%. Defaults to 0 (i.e. 0%).

max_goal
numeric The maximum goal (inclusive) shown for the feature. Note that goal values are specified as proportions, such that a value of 0.1 corresponds to 10%. Defaults to 1 (i.e. 100%).

step_goal
numeric The minimum increment for setting goals. Note that goal are specified as proportions, such that a value of 0.01 corresponds to 1%. Defaults to 0.01 (i.e. 1%).

limit_goal
numeric The minimum goal (inclusive) that can be selected for the feature. Note that goal values are specified as proportions, such that a value of 0.1 corresponds to 10%. Defaults to 0 (i.e. 0%).

current_label
character The display label for the current level of for the feature. Defaults to "Current".

icon
character or shiny.tag Icon to display for the feature This icon should indicate the type of data that underpin the feature. If the argument to icon is a character, it is used with shiny::icon() to generate an shiny.tag icon. Defaults to "map-marked-alt".

id
character unique identifier. Defaults to a random identifier (uuid::UUIDgenerate()).

Value

A Feature object.

Examples

# create new layer
l <- new_layer(source = tempfile(), current = 0.1, total = 12, units = "ha")

# create feature using the layer
f <- new_feature(name = "Intact Alvar", layer = l)

# print object
print(f)

#> <Feature>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     current_label: Current
#>     get_goal: function () 
#>     get_status: function () 
#>     goal: 0.1
#>     icon: shiny.tag
#>     id: ebfc2116-89a4-482a-b36b-af75e371059f
#>     initial_goal: 0.1
#>     initial_status: TRUE
#>     initialize: function (id, name, layer, initial_status, initial_goal, min_goal, 
#>     layer: Layer, R6
#>     limit_goal: 0.1
#>     max_goal: 1
#>     min_goal: 0
#>     name: Intact Alvar
#>     set_goal: function (value) 
#>     set_status: function (value) 
#>     status: TRUE
#>     step_goal: 0.01