New single theme

Create a new SingleTheme object.

new_single_theme(
  name,
  feature,
  initial_status = TRUE,
  round = TRUE,
  icon = "map-marked-alt",
  id = uuid::UUIDgenerate()
)

Arguments

name
character Name to display.

feature
Feature object. For compatibility with new_multi_theme(), a list object can also be supplied that contains a single Feature object.

initial_status
logical The initial status. This is used to display information on whether the theme is selected (or not) for subsequent analysis. Defaults to TRUE

round
logical should all numbers be rounded to the nearest integer? Defaults to TRUE.

icon
shiny.tag Icon to display for the feature This icon should indicate the type of data that underpin the feature. Alternatively, the argument can be a character to automactially generate a shiny.tag icon (using shiny::icon()). Defaults to "map-marked-alt".

id
character unique identifier. Defaults to a random identifier (uuid::UUIDgenerate()).

Value

A SingleTheme object.

Examples

# create new layer
l <- new_layer(source = tempfile(), current = 0.1, total = 12, units = "ha")

# create feature using the layer
f <- new_feature(name = "Intact Alvar Occurrence", layer = l)

# create a theme using the single feature
st <- new_single_theme(name = "Inact Alvar", feature = f)

# print object
print(st)

#> <SingleTheme>
#>   Inherits from: <Theme>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     feature: list
#>     get_feature_goal: function () 
#>     get_feature_status: function () 
#>     get_map_manager_widget_data: function () 
#>     get_parameter: function (name) 
#>     get_solution_settings_widget_data: function () 
#>     icon: shiny.tag
#>     id: cc86ef72-e8d0-4884-8899-8a1b3260b57f
#>     initial_status: TRUE
#>     initialize: function (id, name, feature, initial_status, round, icon) 
#>     name: Inact Alvar
#>     round: TRUE
#>     set_feature_goal: function (value) 
#>     set_feature_status: function (value) 
#>     set_parameter: function (name, value)