Read a configuration file

Read a configuration file from disk. The configuration file will contain all the data and settings related to a set of Theme, Weight, and Include objects.

read_configuration_file(x, spatial_path, attribute_path, boundary_path)

Arguments

x
character file path for configuration file.

spatial_path
character file path for the spatial data.

attribute_path
character file path for the attribute data.

boundary_path
character file path for the attribute data.

Value

A list containing the following elements:

name

A character value indicating the name.

dataset

A Dataset object.

themes

A list of Theme objects.

weights

A list of Weight objects.

includes

A list of Include objects.

mode

A character value indicating the mode.

Examples

# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_data.yaml", package = "locationmisc")
f2 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f4 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# read configuration file
x <- read_configuration_file(f1, f2, f3, f4)

# print results
print(x)
#> $name
#> [1] "Example GeoTIFF dataset"
#> 
#> $dataset
#> Dataset
#>   paths:
#>     spatial: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_spatial.tif
#>     attribute: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_attribute.csv.gz
#>     boundary: /tmp/Rtmpo56wvS/temp_libpath193741748187a/locationmisc/extdata/sim_raster_boundary.csv.gz
#> 
#> $themes
#> $themes[[1]]
#> Theme
#>   id:        f2aec00c-2eca-4444-94eb-703bf13650d9
#>   name:      Olive-headed Lorikeet
#>   mandatory: FALSE
#>   feature: 
#>     Olive-headed Lorikeet habitat [status: TRUE, current: 0.72, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Olive_headed_Lorikeet [total: 1052 ha]
#> 
#> $themes[[2]]
#> Theme
#>   id:        97c68c68-ca9c-4d24-af04-741e8eb1bbe0
#>   name:      Purple Honeycreeper
#>   mandatory: FALSE
#>   feature: 
#>     Purple Honeycreeper habitat [status: TRUE, current: 0.74, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Purple_Honeycreeper [total: 1667 ha]
#> 
#> $themes[[3]]
#> Theme
#>   id:        7afa6a70-7980-4f11-be63-c036ba391c26
#>   name:      Least Boobook
#>   mandatory: TRUE
#>   feature: 
#>     Least Boobook habitat [status: TRUE, current: 0.72, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Least_Boobook [total: 1678 ha]
#> 
#> $themes[[4]]
#> Theme
#>   id:        9c801378-cd12-468a-ac21-91e0af7533fe
#>   name:      Penduline-Tits
#>   mandatory: FALSE
#>   feature: 
#>     African Penduline-Tit [status: TRUE, current: 0.73, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#African_Penduline_Tit [total: 1097 ha]
#>     Black-headed Penduline-Tit [status: TRUE, current: 0.72, goal: 0.5]
#>       variable: .../sim_raster_spatial.tif#Black_headed_Penduline_Tit [total: 433 ha]
#>     Verdin [status: TRUE, current: 0.75, goal: 0.89]
#>       variable: .../sim_raster_spatial.tif#Verdin [total: 443 ha]
#>     African Penduline-Tit (Buff-bellied) [status: TRUE, current: 0.71, goal: 0.58]
#>       variable: .../sim_raster_spatial.tif#African_Penduline_Tit_Buff_bellied_ [total: 1800 ha]
#>     Sennar Penduline-Tit [status: TRUE, current: 0.74, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#Sennar_Penduline_Tit [total: 1075 ha]
#> 
#> $themes[[5]]
#> Theme
#>   id:        fccf1219-a8c1-4e0a-a6cf-a0b9c1ff64b2
#>   name:      Bowerbirds
#>   mandatory: FALSE
#>   feature: 
#>     Streaked Bowerbird [status: TRUE, current: 0.71, goal: 0.68]
#>       variable: .../sim_raster_spatial.tif#Streaked_Bowerbird [total: 609 ha]
#>     Black-capped Catbird [status: TRUE, current: 0.74, goal: 0.88]
#>       variable: .../sim_raster_spatial.tif#Black_capped_Catbird [total: 1258 ha]
#>     Golden-fronted Bowerbird [status: TRUE, current: 0.72, goal: 0.62]
#>       variable: .../sim_raster_spatial.tif#Golden_fronted_Bowerbird [total: 1188 ha]
#>     Fawn-breasted Bowerbird [status: TRUE, current: 0.75, goal: 0.85]
#>       variable: .../sim_raster_spatial.tif#Fawn_breasted_Bowerbird [total: 691 ha]
#> 
#> 
#> $weights
#> $weights[[1]]
#> Weight
#>   id:       6171bc33-7d6d-4d9c-9a4c-86c803b5e05e
#>   name:     Population density
#>   variable: .../sim_raster_spatial.tif#Population_density [total: 149.91 ha]
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> $weights[[2]]
#> Weight
#>   id:       becc0e55-e498-4f58-b045-56facaf5bb39
#>   name:     Invasive species
#>   variable: .../sim_raster_spatial.tif#Invasive_species [total: 149.98 ha]
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> $weights[[3]]
#> Weight
#>   id:       43fa6e37-c3ef-4cd0-b78d-8a088461f01e
#>   name:     Archaeological site
#>   variable: .../sim_raster_spatial.tif#Archaeological_site [total: 144.76 ha]
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> 
#> $includes
#> $includes[[1]]
#> Include
#>   id:       12c1fd44-62b6-48c6-b185-cbf356b9e3d2
#>   name:     Special Protection Area
#>   variable: .../sim_raster_spatial.tif#Special_Protection_Area [total: 156 ]
#>   visible:  TRUE
#>   status:   TRUE
#> 
#> $includes[[2]]
#> Include
#>   id:       65901d7a-8fba-46c3-83cc-ec8d10c4b960
#>   name:     Site of Community Importance
#>   variable: .../sim_raster_spatial.tif#Site_of_Community_Importance [total: 138 ]
#>   visible:  TRUE
#>   status:   TRUE
#> 
#> 
#> $mode
#> [1] "advanced"
#>