Simulate a solution settings object

This function simulates a solution settings object. It is primarily used for testing the package.

simulate_solution_settings(
  n_single_themes = 5,
  n_multi_themes = 15,
  n_weights = 5,
  lambda = 5
)

Arguments

n_single_themes
integer number of themes with a single feature to simulate. Defaults to 5.

n_multi_themes
integer number of themes with a multiple features to simulate. Defaults to 15.

n_weights
integer number of weights to simulate. Defaults to 5.

lambda
numeric lambda parameter for simulating the number of features for each theme. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A SolutionSettings object.

See also

Examples

# simulate solution settings
x <- simulate_solution_settings(
  n_single_themes = 3, n_multi_themes = 2, n_weights = 1)

# print results
print(x)
#> SolutionSettings
#>   themes: 
#>     Eurasian Jay:
#>       Eurasian Jay habitat [status: TRUE, current: 0.48, goal: 0.88]
#>         dataset: .../file57cc12513ed5 [total: 596.99 ha]
#>     Brown-headed Apalis:
#>       Brown-headed Apalis habitat [status: TRUE, current: 0.2, goal: 0.77]
#>         dataset: .../file57cc6206a96f [total: 1049.67 ha]
#>     Bronzed Cowbird (Bronzed):
#>       Bronzed Cowbird (Bronzed) habitat [status: TRUE, current: 0.38, goal: 0.81]
#>         dataset: .../file57cc1db8949 [total: 3932.25 ha]
#>     Gulls, Terns, and Skimmers:
#>       Iceland Gull (glaucoides) [status: TRUE, current: 0.2, goal: 0.59]
#>         dataset: .../file57cc7e5d5191 [total: 1030.9 ha]
#>       Belcher's Gull [status: TRUE, current: 0.31, goal: 0.51]
#>         dataset: .../file57cc10bb0569 [total: 8292.26 ha]
#>       Mew Gull (Kamchatka) [status: TRUE, current: 0.12, goal: 0.57]
#>         dataset: .../file57cc1feea913 [total: 7218.24 ha]
#>       Herring Gull [status: TRUE, current: 0.5, goal: 0.57]
#>         dataset: .../file57cc6a154dc9 [total: 7880.92 ha]
#>     Jacamars:
#>       White-eared Jacamar [status: TRUE, current: 0.49, goal: 0.83]
#>         dataset: .../file57cc61450ec0 [total: 1733.61 ha]
#>       Rufous-tailed Jacamar (Black-chinned) [status: TRUE, current: 0.29, goal: 0.54]
#>         dataset: .../file57cc52738e93 [total: 8513.84 ha]
#>       Purus Jacamar [status: TRUE, current: 0.12, goal: 0.55]
#>         dataset: .../file57cc173ca69a [total: 9866.44 ha]
#>       Paradise Jacamar [status: TRUE, current: 0.53, goal: 0.9]
#>         dataset: .../file57cc33c9ed06 [total: 9371.23 ha]
#>       Coppery-chested Jacamar [status: TRUE, current: 0.24, goal: 0.58]
#>         dataset: .../file57cc70d5b6c1 [total: 8746.83 ha]
#>       White-throated Jacamar [status: TRUE, current: 0.54, goal: 0.76]
#>         dataset: .../file57cc74c03252 [total: 9615.05 ha]
#>   weights: 
#>     Archaeological site [status: TRUE, factor: 0
#>       dataset: .../file57cc26bdbf7f [total: 7584.81 ???]]