Simulate themes

This function simulates SingleTheme and MultiTheme objects.

simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for each MultiTheme object. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "sim_raster_spatial.tif", package = "locationmisc")
f2 <- system.file(
 "extdata", "sim_raster_attribute.csv.gz", package = "locationmisc")
f3 <- system.file(
 "extdata", "sim_raster_boundary.csv.gz", package = "locationmisc")

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
#> .....
#> ....

# print results
print(x)
#> [[1]]
#> Theme
#>   id:        6bab3a2e-b1a4-402d-8a3a-0055598ef727
#>   name:      Blue-headed Parrot (Blue-headed)
#>   mandatory: TRUE
#>   feature: 
#>     Blue-headed Parrot (Blue-headed) habitat [status: TRUE, current: 0.2, goal: 0.6]
#>       variable: .../sim_raster_spatial.tif#Blue_headed_Parrot_Blue_headed_ [total: 2574.85 ha]
#> 
#> [[2]]
#> Theme
#>   id:        16aeceb1-acdc-4386-b2ec-ed8bd6cb3483
#>   name:      Chestnut-headed Bee-eater
#>   mandatory: TRUE
#>   feature: 
#>     Chestnut-headed Bee-eater habitat [status: TRUE, current: 0.39, goal: 0.67]
#>       variable: .../sim_raster_spatial.tif#Chestnut_headed_Bee_eater [total: 1291 ha]
#> 
#> [[3]]
#> Theme
#>   id:        f28e5f40-9872-4bd5-bcb3-05d537951d8f
#>   name:      Moustached Flowerpiercer (mystacalis)
#>   mandatory: TRUE
#>   feature: 
#>     Moustached Flowerpiercer (mystacalis) habitat [status: TRUE, current: 0.26, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Moustached_Flowerpiercer_mystacalis_ [total: 478 ha]
#> 
#> [[4]]
#> Theme
#>   id:        e102ed74-0668-4384-90ec-81997160dd95
#>   name:      Northern Storm-Petrels
#>   mandatory: FALSE
#>   feature: 
#>     Ringed Storm-Petrel [status: TRUE, current: 0.18, goal: 0.61]
#>       variable: .../sim_raster_spatial.tif#Ringed_Storm_Petrel [total: 5152.8 ha]
#>     Ainley's Storm-Petrel [status: TRUE, current: 0.49, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Ainley_s_Storm_Petrel [total: 11860.98 ha]
#>     Guadalupe Storm-Petrel [status: TRUE, current: 0.36, goal: 0.59]
#>       variable: .../sim_raster_spatial.tif#Guadalupe_Storm_Petrel [total: 5057.88 ha]
#>     Matsudaira's Storm-Petrel [status: TRUE, current: 0.22, goal: 0.69]
#>       variable: .../sim_raster_spatial.tif#Matsudaira_s_Storm_Petrel [total: 226.66 ha]
#>     Band-rumped Storm-Petrel [status: TRUE, current: 0.54, goal: 0.86]
#>       variable: .../sim_raster_spatial.tif#Band_rumped_Storm_Petrel [total: 4647.41 ha]
#> 
#> [[5]]
#> Theme
#>   id:        9f5425bd-447e-4d99-9612-301fab93c132
#>   name:      Todies
#>   mandatory: FALSE
#>   feature: 
#>     Puerto Rican Tody [status: TRUE, current: 0.42, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#Puerto_Rican_Tody [total: 11569.62 ha]
#>     Narrow-billed Tody [status: TRUE, current: 0.55, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Narrow_billed_Tody [total: 18687.69 ha]
#>     Cuban Tody [status: TRUE, current: 0.24, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Cuban_Tody [total: 18246.34 ha]
#>     Jamaican Tody [status: TRUE, current: 0.46, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Jamaican_Tody [total: 5478.13 ha]
#>